Moved the TokenDisplay component to its own file
