+++ components/index.ts
@@ -1,2 +1,3 @@
 export { Connect } from './Connect'
 export { Erc1155Balance } from './Erc1155Balance'
+export { TokenDisplay } from './TokenDisplay'
